payload/requirements.txt
pyserial

pillow

typing